added sound

+++ src/kmath.rs
@@ -1,27 +1,27 @@
-pub use std::f64::consts::PI;
-pub static DEG_TO_RAD: f64 = 180.0 / PI;
+pub use std::f32::consts::PI;
+pub static DEG_TO_RAD: f32 = 180.0 / PI;
 
 /***************************************************
  * Easing
  ***************************************************/
-pub fn lerp(x1: f64, x2: f64, t: f64) -> f64 {
+pub fn lerp(x1: f32, x2: f32, t: f32) -> f32 {
     x1 * (1.0 - t) + x2 * t
 }
 
-pub fn unlerp(x: f64, t1: f64, t2: f64) -> f64 {
+pub fn unlerp(x: f32, t1: f32, t2: f32) -> f32 {
     (x - t1) / (t2 - t1)
 }
 
-pub fn remap(x: f64, from_low: f64, from_high: f64, to_low: f64, to_high: f64) -> f64 {
+pub fn remap(x: f32, from_low: f32, from_high: f32, to_low: f32, to_high: f32) -> f32 {
     lerp(to_low, to_high, unlerp(x, from_low, from_high))
 }
 
-pub fn cubic_bezier(start: Vec2, c1: Vec2, c2: Vec2, end: Vec2, t: f64) -> Vec2 {
+pub fn cubic_bezier(start: Vec2, c1: Vec2, c2: Vec2, end: Vec2, t: f32) -> Vec2 {
     start.lerp(c1.lerp(c2.lerp(end, t), t), t)
 }
 
 // t 0..1
-pub fn smoothstep(t: f64) -> f64 {
+pub fn smoothstep(t: f32) -> f32 {
     t * t * (3. - 2. * t)
 }
 
@@ -29,7 +29,7 @@ pub fn smoothstep(t: f64) -> f64 {
  * RNG
  ***************************************************/
 
- pub fn floorfrac(x: f64) -> (f64, f64) {
+ pub fn floorfrac(x: f32) -> (f32, f32) {
     let floor = x.floor();
     if x < 0.0 {
         (floor, (floor - x).abs())
@@ -49,25 +49,25 @@ pub fn khash2i(x: i32, y: i32, seed: u32) -> u32 {
     khash((x as u32).wrapping_mul(123176957).wrapping_add((y as u32).wrapping_mul(489172373)).wrapping_add(seed))
 }
 
-pub fn krand(seed: u32) -> f64 {
-    khash(seed) as f64 / 4294967295.0
+pub fn krand(seed: u32) -> f32 {
+    khash(seed) as f32 / 4294967295.0
 }
 
-pub fn kuniform(seed: u32, min: f64, max: f64) -> f64 {
-    min + (khash(seed) as f64 / 4294967295.0) * (max - min)
+pub fn kuniform(seed: u32, min: f32, max: f32) -> f32 {
+    min + (khash(seed) as f32 / 4294967295.0) * (max - min)
 }
 
-pub fn chance(seed: u32, percent: f64) -> bool {
+pub fn chance(seed: u32, percent: f32) -> bool {
     krand(seed) < percent
 }
 
-pub fn noise1d(t: f64, seed: u32) -> f64 {
+pub fn noise1d(t: f32, seed: u32) -> f32 {
     let hstart = kuniform(seed + 489172373 * t.floor() as u32, 0.0, 1.0);
     let hend = kuniform(seed + 489172373 * (t.floor() + 1.0) as u32, 0.0, 1.0);
     lerp(hstart, hend, smoothstep(t.fract()))
 }
 
-pub fn noise2d(x: f64, y: f64, seed: u32) -> f64 {
+pub fn noise2d(x: f32, y: f32, seed: u32) -> f32 {
     let (xfloor, xfrac) = floorfrac(x);
     let (yfloor, yfrac) = floorfrac(y);
 
@@ -98,31 +98,31 @@ pub fn noise2d(x: f64, y: f64, seed: u32) -> f64 {
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Vec2 {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vec2 {
-    pub const fn new(x: f64, y: f64) -> Vec2 { Vec2{x, y} }
+    pub const fn new(x: f32, y: f32) -> Vec2 { Vec2{x, y} }
     pub fn plus(&self, other: Vec2) -> Vec2 { {Vec2::new(self.x + other.x, self.y + other.y)} }
     pub fn minus(&self, other: Vec2) -> Vec2 { {Vec2::new(self.x + other.x, self.y + other.y)} }
-    pub fn new_r_theta(r: f64, theta: f64) -> Vec2 { Vec2{x: r * theta.cos(), y: r * theta.sin()} }
-    pub fn mul_scalar(&self, scalar: f64) -> Vec2 { Vec2::new(self.x * scalar, self.y * scalar) }
-    pub fn div_scalar(&self, scalar: f64) -> Vec2 { Vec2::new(self.x / scalar, self.y / scalar) }
-    pub fn magnitude(&self) -> f64 { (self.x*self.x + self.y*self.y).sqrt() }
-    pub fn dist(&self, other: Vec2) -> f64 { (*self - other).magnitude() }
+    pub fn new_r_theta(r: f32, theta: f32) -> Vec2 { Vec2{x: r * theta.cos(), y: r * theta.sin()} }
+    pub fn mul_scalar(&self, scalar: f32) -> Vec2 { Vec2::new(self.x * scalar, self.y * scalar) }
+    pub fn div_scalar(&self, scalar: f32) -> Vec2 { Vec2::new(self.x / scalar, self.y / scalar) }
+    pub fn magnitude(&self) -> f32 { (self.x*self.x + self.y*self.y).sqrt() }
+    pub fn dist(&self, other: Vec2) -> f32 { (*self - other).magnitude() }
     pub fn normalize(&self) -> Vec2 { let m = self.magnitude(); if m == 0.0 { *self } else { self.div_scalar(self.magnitude()) }}
-    pub fn lerp(&self, other: Vec2, t: f64) -> Vec2 { Vec2::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t)) }
-    pub fn rotate(&self, radians: f64) -> Vec2 { 
+    pub fn lerp(&self, other: Vec2, t: f32) -> Vec2 { Vec2::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t)) }
+    pub fn rotate(&self, radians: f32) -> Vec2 { 
         Vec2::new(
             self.x * radians.cos() - self.y * radians.sin(), 
             self.x * radians.sin() + self.y * radians.cos()
         ) 
     }
-    pub fn offset_r_theta(&self, r: f64, theta: f64) -> Vec2 {
+    pub fn offset_r_theta(&self, r: f32, theta: f32) -> Vec2 {
         *self + Vec2::new(r, 0.0).rotate(theta)
     }
-    pub fn promote(&self, z: f64) -> Vec3 {
+    pub fn promote(&self, z: f32) -> Vec3 {
         Vec3::new(self.x, self.y, z)
     }
     pub fn transform(&self, from: Rect, to: Rect) -> Vec2 {
@@ -150,7 +150,7 @@ impl Vec2 {
 
         Vec2::new(a*c + b*d, b*c - a*d) / denom
     }
-    pub fn rect_centered(&self, w: f64, h: f64) -> Rect {
+    pub fn rect_centered(&self, w: f32, h: f32) -> Rect {
         Rect::new(self.x - w/2.0, self.y - h/2.0, w, h)
     }
     
@@ -172,15 +172,15 @@ impl std::ops::Add<Vec2> for Vec2 {
     }
 }
 
-impl std::ops::Mul<f64> for Vec2 {
+impl std::ops::Mul<f32> for Vec2 {
     type Output = Vec2;
 
-    fn mul(self, _rhs: f64) -> Vec2 {
+    fn mul(self, _rhs: f32) -> Vec2 {
         self.mul_scalar(_rhs)
     }
 }
 
-impl std::ops::Mul<Vec2> for f64 {
+impl std::ops::Mul<Vec2> for f32 {
     type Output = Vec2;
 
     fn mul(self, _rhs: Vec2) -> Vec2 {
@@ -188,10 +188,10 @@ impl std::ops::Mul<Vec2> for f64 {
     }
 }
 
-impl std::ops::Div<f64> for Vec2 {
+impl std::ops::Div<f32> for Vec2 {
     type Output = Vec2;
 
-    fn div(self, _rhs: f64) -> Vec2 {
+    fn div(self, _rhs: f32) -> Vec2 {
         self.div_scalar(_rhs)
     }
 }
@@ -206,21 +206,21 @@ impl std::ops::Neg for Vec2 {
 
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub struct Vec3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl Vec3 {
-    pub const fn new(x: f64, y: f64, z: f64) -> Vec3 { Vec3{x, y, z} }
-    pub fn mul_scalar(&self, scalar: f64) -> Vec3 { Vec3::new(self.x * scalar, self.y * scalar, self.z * scalar) }
-    pub fn div_scalar(&self, scalar: f64) -> Vec3 { Vec3::new(self.x / scalar, self.y / scalar, self.z / scalar) }
-    pub fn magnitude(&self) -> f64 { (self.x*self.x + self.y*self.y + self.z*self.z).sqrt() }
-    pub fn square_distance(&self) -> f64 { self.x*self.x + self.y*self.y + self.z*self.z }
+    pub const fn new(x: f32, y: f32, z: f32) -> Vec3 { Vec3{x, y, z} }
+    pub fn mul_scalar(&self, scalar: f32) -> Vec3 { Vec3::new(self.x * scalar, self.y * scalar, self.z * scalar) }
+    pub fn div_scalar(&self, scalar: f32) -> Vec3 { Vec3::new(self.x / scalar, self.y / scalar, self.z / scalar) }
+    pub fn magnitude(&self) -> f32 { (self.x*self.x + self.y*self.y + self.z*self.z).sqrt() }
+    pub fn square_distance(&self) -> f32 { self.x*self.x + self.y*self.y + self.z*self.z }
     pub fn normalize(&self) -> Vec3 { self.div_scalar(self.magnitude()) }
-    pub fn lerp(&self, other: Vec3, t: f64) -> Vec3 { Vec3::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t), self.z*(1.0-t) + other.z*(t)) }
-    pub fn dist(&self, other: Vec3) -> f64 {(*self - other).magnitude().sqrt()}
-    pub fn dot(&self, other: Vec3) -> f64 {self.x*other.x + self.y*other.y + self.z*other.z} // is squ dist lol
+    pub fn lerp(&self, other: Vec3, t: f32) -> Vec3 { Vec3::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t), self.z*(1.0-t) + other.z*(t)) }
+    pub fn dist(&self, other: Vec3) -> f32 {(*self - other).magnitude().sqrt()}
+    pub fn dot(&self, other: Vec3) -> f32 {self.x*other.x + self.y*other.y + self.z*other.z} // is squ dist lol
     pub fn cross(&self, other: Vec3) -> Vec3 {
         Vec3::new(
             self.y*other.z - self.z*other.y,
@@ -228,10 +228,10 @@ impl Vec3 {
             self.x*other.y - self.y*other.x,
         )
     }
-    pub fn rotate_about_vec3(&self, axis: Vec3, theta: f64) -> Vec3 {
+    pub fn rotate_about_vec3(&self, axis: Vec3, theta: f32) -> Vec3 {
         *self*theta.cos() + (axis.cross(*self)*theta.sin()) + axis * (axis.dot(*self)*(1.0 - theta.cos()))
     }
-    pub fn promote(&self, w: f64) -> Vec4 {
+    pub fn promote(&self, w: f32) -> Vec4 {
         Vec4::new(self.x, self.y, self.z, w)
     }
 }
@@ -252,15 +252,15 @@ impl std::ops::Add<Vec3> for Vec3 {
     }
 }
 
-impl std::ops::Mul<f64> for Vec3 {
+impl std::ops::Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, _rhs: f64) -> Vec3 {
+    fn mul(self, _rhs: f32) -> Vec3 {
         self.mul_scalar(_rhs)
     }
 }
 
-impl std::ops::Mul<Vec3> for f64 {
+impl std::ops::Mul<Vec3> for f32 {
     type Output = Vec3;
 
     fn mul(self, _rhs: Vec3) -> Vec3 {
@@ -268,10 +268,10 @@ impl std::ops::Mul<Vec3> for f64 {
     }
 }
 
-impl std::ops::Div<f64> for Vec3 {
+impl std::ops::Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, _rhs: f64) -> Vec3 {
+    fn div(self, _rhs: f32) -> Vec3 {
         self.div_scalar(_rhs)
     }
 }
@@ -303,27 +303,27 @@ impl std::fmt::Display for Vec3 {
 
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub struct Vec4 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-    pub w: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
+    pub w: f32,
 }
 
 impl Vec4 {
-    pub const fn new(x: f64, y: f64, z: f64, w: f64) -> Vec4 { Vec4{x, y, z, w} }
-    pub fn mul_scalar(&self, scalar: f64) -> Vec4 { Vec4::new(self.x * scalar, self.y * scalar, self.z * scalar, self.w * scalar) }
-    pub fn div_scalar(&self, scalar: f64) -> Vec4 { Vec4::new(self.x / scalar, self.y / scalar, self.z / scalar, self.w / scalar) }
-    pub fn magnitude(&self) -> f64 { (self.x*self.x + self.y*self.y + self.z*self.z + self.w*self.w).sqrt() }
-    pub fn square_distance(&self) -> f64 { self.x*self.x + self.y*self.y + self.z*self.z + self.w*self.w }
+    pub const fn new(x: f32, y: f32, z: f32, w: f32) -> Vec4 { Vec4{x, y, z, w} }
+    pub fn mul_scalar(&self, scalar: f32) -> Vec4 { Vec4::new(self.x * scalar, self.y * scalar, self.z * scalar, self.w * scalar) }
+    pub fn div_scalar(&self, scalar: f32) -> Vec4 { Vec4::new(self.x / scalar, self.y / scalar, self.z / scalar, self.w / scalar) }
+    pub fn magnitude(&self) -> f32 { (self.x*self.x + self.y*self.y + self.z*self.z + self.w*self.w).sqrt() }
+    pub fn square_distance(&self) -> f32 { self.x*self.x + self.y*self.y + self.z*self.z + self.w*self.w }
     pub fn normalize(&self) -> Vec4 { self.div_scalar(self.magnitude()) }
-    pub fn lerp(&self, other: Vec4, t: f64) -> Vec4 { Vec4::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t), self.z*(1.0-t) + other.z*(t), self.w*(1.0-t) + other.w*t) }
-    pub fn dist(&self, other: Vec4) -> f64 {(*self - other).magnitude().sqrt()}
-    pub fn dot(&self, other: Vec4) -> f64 {self.x*other.x + self.y*other.y + self.z*other.z + self.w*other.w} // is squ dist lol
+    pub fn lerp(&self, other: Vec4, t: f32) -> Vec4 { Vec4::new(self.x*(1.0-t) + other.x*(t), self.y*(1.0-t) + other.y*(t), self.z*(1.0-t) + other.z*(t), self.w*(1.0-t) + other.w*t) }
+    pub fn dist(&self, other: Vec4) -> f32 {(*self - other).magnitude().sqrt()}
+    pub fn dot(&self, other: Vec4) -> f32 {self.x*other.x + self.y*other.y + self.z*other.z + self.w*other.w} // is squ dist lol
     pub fn hsv_to_rgb(&self) -> Vec4 {
         let v = self.z;
         let hh = (self.x % 360.0) / 60.0;
         let i = hh.floor() as i32;
-        let ff = hh - i as f64;
+        let ff = hh - i as f32;
         let p = self.z * (1.0 - self.y);
         let q = self.z * (1.0 - self.y * ff);
         let t = self.z * (1.0 - self.y * (1.0 - ff));
@@ -339,11 +339,11 @@ impl Vec4 {
     }
 
     // h in degrees
-    pub fn from_hsv(h: f64, s: f64, v: f64, w: f64) -> Vec4 {
+    pub fn from_hsv(h: f32, s: f32, v: f32, w: f32) -> Vec4 {
         Vec4::new(h, s, v, w).hsv_to_rgb()
     }
 
-    pub fn grey(v: f64) -> Vec4 {
+    pub fn grey(v: f32) -> Vec4 {
         Vec4::new(v, v, v, 1.0)
     }
 }
@@ -364,15 +364,15 @@ impl std::ops::Add<Vec4> for Vec4 {
     }
 }
 
-impl std::ops::Mul<f64> for Vec4 {
+impl std::ops::Mul<f32> for Vec4 {
     type Output = Vec4;
 
-    fn mul(self, _rhs: f64) -> Vec4 {
+    fn mul(self, _rhs: f32) -> Vec4 {
         self.mul_scalar(_rhs)
     }
 }
 
-impl std::ops::Mul<Vec4> for f64 {
+impl std::ops::Mul<Vec4> for f32 {
     type Output = Vec4;
 
     fn mul(self, _rhs: Vec4) -> Vec4 {
@@ -380,10 +380,10 @@ impl std::ops::Mul<Vec4> for f64 {
     }
 }
 
-impl std::ops::Div<f64> for Vec4 {
+impl std::ops::Div<f32> for Vec4 {
     type Output = Vec4;
 
-    fn div(self, _rhs: f64) -> Vec4 {
+    fn div(self, _rhs: f32) -> Vec4 {
         self.div_scalar(_rhs)
     }
 }
@@ -420,20 +420,20 @@ impl std::fmt::Display for Vec4 {
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Rect {
-    pub x: f64,
-    pub y: f64,
-    pub w: f64,
-    pub h: f64,
+    pub x: f32,
+    pub y: f32,
+    pub w: f32,
+    pub h: f32,
 }
 
 impl Rect {
-    pub fn new(x: f64, y: f64, w: f64, h: f64) -> Rect {
+    pub fn new(x: f32, y: f32, w: f32, h: f32) -> Rect {
         Rect{x,y,w,h}
     }
-    pub fn centered(p: Vec2, w: f64, h: f64) -> Rect {
+    pub fn centered(p: Vec2, w: f32, h: f32) -> Rect {
         Rect::new(p.x - w/2.0, p.y - h/2.0, w, h)
     }
-    pub fn child(&self, x: f64, y: f64, w: f64, h: f64) -> Rect {
+    pub fn child(&self, x: f32, y: f32, w: f32, h: f32) -> Rect {
         Rect::new(
             self.x + x*self.w,
             self.y + y*self.h,
@@ -442,12 +442,12 @@ impl Rect {
         )
     }
     pub fn grid_child(&self, x: i32, y: i32, w: i32, h: i32) -> Rect {
-        let r_w = self.w / w as f64;
-        let r_h = self.h / h as f64;
+        let r_w = self.w / w as f32;
+        let r_h = self.h / h as f32;
 
         Rect::new(
-            self.x + r_w * x as f64,
-            self.y + r_h * y as f64,
+            self.x + r_w * x as f32,
+            self.y + r_h * y as f32,
             r_w,
             r_h,
         )
@@ -456,7 +456,7 @@ impl Rect {
         let s = self.w.min(self.h);
         Rect::new_centered(self.x + self.w / 2.0, self.y + self.h / 2.0, s, s)
     }
-    pub fn fit_aspect_ratio(&self, a: f64) -> Rect {
+    pub fn fit_aspect_ratio(&self, a: f32) -> Rect {
         let our_a = self.w / self.h;
         if our_a < a {
             // big a means wide
@@ -471,7 +471,7 @@ impl Rect {
             Rect::new(other_x, self.y, other_w, self.h)
         }
     }
-    pub fn fill_aspect_ratio(&self, a: f64) -> Rect {
+    pub fn fill_aspect_ratio(&self, a: f32) -> Rect {
         let our_a = self.w / self.h;
         if our_a < a {
             //wider
@@ -481,7 +481,7 @@ impl Rect {
             Rect::centered(self.centroid(), self.w, self.h * (our_a / a))
         }
     } 
-    pub fn lerp(&self, other: Rect, t: f64) -> Rect {
+    pub fn lerp(&self, other: Rect, t: f32) -> Rect {
         Rect::new(
             lerp(self.x, other.x, t),
             lerp(self.y, other.y, t),
@@ -489,35 +489,35 @@ impl Rect {
             lerp(self.h, other.h, t),
         )
     }
-    pub fn aspect(&self) -> f64 {
+    pub fn aspect(&self) -> f32 {
         self.w / self.h
     }
     pub fn centroid(&self) -> Vec2 {
         Vec2::new(self.x + self.w/2.0, self.y + self.h/2.0)
     }
-    pub fn new_centered(x: f64, y: f64, w: f64, h: f64) -> Rect {
+    pub fn new_centered(x: f32, y: f32, w: f32, h: f32) -> Rect {
         Rect::new(x-w/2.0, y-h/2.0, w, h)
     }
     pub fn translate(&self, v: Vec2) -> Rect {
         return Rect::new(self.x + v.x, self.y + v.y, self.w, self.h);
     }
-    pub fn dilate(&self, d: f64) -> Rect {
+    pub fn dilate(&self, d: f32) -> Rect {
         return Rect::new(self.x - d, self.y - d, self.w + 2.0*d, self.h + 2.0*d);
     }
-    pub fn dilate_pc(&self, percent: f64) -> Rect {
+    pub fn dilate_pc(&self, percent: f32) -> Rect {
         let amount = self.w.min(self.h) * percent;
         self.dilate(amount)
     }
-    pub fn left(self) -> f64 {
+    pub fn left(self) -> f32 {
         self.x
     }
-    pub fn right(self) -> f64 {
+    pub fn right(self) -> f32 {
         self.x + self.w
     }
-    pub fn top(self) -> f64 {
+    pub fn top(self) -> f32 {
         self.y
     }
-    pub fn bot(self) -> f64 {
+    pub fn bot(self) -> f32 {
         self.y + self.h
     }
     pub fn tl(self) -> Vec2 {
@@ -540,7 +540,7 @@ impl Rect {
         Vec2::new((point.x - self.x) / self.w, (point.y - self.y) / self.h)
     }
     pub fn grid_square(self, point: Vec2, w: i32, h: i32) -> (i32, i32) {
-        ((w as f64 * point.x) as i32, (h as f64 * point.y) as i32)
+        ((w as f32 * point.x) as i32, (h as f32 * point.y) as i32)
     }
     pub fn snap(&self, mut point: Vec2) -> Vec2 {
         if point.x < self.x {
@@ -568,7 +568,7 @@ impl Rect {
     }
 
     // 5 cases: both a in b, both b in a, a left in b, b left in a, no overlap
-    fn overlap_amount(a1: f64, a2: f64, b1: f64, b2: f64) -> f64 {
+    fn overlap_amount(a1: f32, a2: f32, b1: f32, b2: f32) -> f32 {
         let a1_in_b = a1 >= b1 && a1 <= b2;
         let a2_in_b = a2 >= b1 && a2 <= b2;
         let b1_in_a = b1 >= a1 && b1 <= a2;
@@ -605,10 +605,10 @@ impl Rect {
         )
     }
 
-    pub fn split_ud(&self, t: f64) -> (Rect, Rect) {
+    pub fn split_ud(&self, t: f32) -> (Rect, Rect) {
         (self.child(0.0, 0.0, 1.0, t), self.child(0.0, t, 1.0, 1.0 - t))
     }
-    pub fn split_lr(&self, t: f64) -> (Rect, Rect) {
+    pub fn split_lr(&self, t: f32) -> (Rect, Rect) {
         (self.child(0.0, 0.0, t, 1.0), self.child(t, 0.0, 1.0 - t, 1.0))
     }
 
@@ -629,7 +629,7 @@ impl Triangle {
         Triangle {a, b, c}
     }
 
-    pub fn dilate(&self, d: f64) -> Triangle {
+    pub fn dilate(&self, d: f32) -> Triangle {
         let centroid = Vec2::new((self.a.x + self.b.x + self.c.x) / 3.0, (self.a.y + self.b.y + self.c.y) / 3.0);
         Triangle::new(
             self.a + (self.a - centroid) * d,
@@ -661,7 +661,7 @@ pub fn mat4_mul(a: [f32; 16], b: [f32; 16]) -> [f32; 16] {
     let mut r = [0.0f32; 16];
     for i in 0..4 {
         for j in 0..4 {
-            r[4* j + i] = Vec4::new(a[4*j + 0] as f64, a[4*j + 1] as f64, a[4*j + 2] as f64, a[4*j + 3] as f64).dot(Vec4::new(b[4*0 + i] as f64, b[4*1 + i] as f64, b[4*2 + i] as f64, b[4*3 + i] as f64)) as f32;
+            r[4* j + i] = Vec4::new(a[4*j + 0] as f32, a[4*j + 1] as f32, a[4*j + 2] as f32, a[4*j + 3] as f32).dot(Vec4::new(b[4*0 + i] as f32, b[4*1 + i] as f32, b[4*2 + i] as f32, b[4*3 + i] as f32)) as f32;
         }
     }
     r
